fixup! Add docker compose config
