+++ backend/bracket/routes/auth.py
@@ -12,9 +12,9 @@ from bracket.config import config
 from bracket.database import database
 from bracket.models.db.tournament import Tournament
 from bracket.models.db.user import UserInDB, UserPublic
-from bracket.schema import tournaments, users
-from bracket.sql.users import get_user_access_to_club, get_user_access_to_tournament
-from bracket.utils.db import fetch_all_parsed, fetch_one_parsed
+from bracket.schema import tournaments
+from bracket.sql.users import get_user, get_user_access_to_club, get_user_access_to_tournament
+from bracket.utils.db import fetch_all_parsed
 from bracket.utils.security import pwd_context
 from bracket.utils.types import assert_some
 
@@ -53,10 +53,6 @@ class TokenData(BaseModel):
 
 def verify_password(plain_password: str, hashed_password: str) -> bool:
     return pwd_context.verify(plain_password, hashed_password)
-
-
-async def get_user(email: str) -> UserInDB | None:
-    return await fetch_one_parsed(database, UserInDB, users.select().where(users.c.email == email))
 
 
 async def authenticate_user(email: str, password: str) -> UserInDB | None:
